fix(server): Implement MCP protocol negotiation. (#341)
According to [the MCP spec](https://modelcontextprotocol.io/specification/2025-03-26/basic/lifecycle#version-negotiation):

> If the server supports the requested protocol version, it MUST respond with
> the same version. Otherwise, the server MUST respond with another protocol
> version it supports.

This adds `mcp.ValidProtocolVersions`, a slice with all currently specified MCP
versions.

In the `server` package we check if the client provided a known MCP version
and, if so, return that version. Otherwise `LATEST_PROTOCOL_VERSION` is
returned as previously.